+++ CHANGELOG.md
@@ -9,11 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Modified zenodo.json file to include more Orcid Ids
   [CATS-884](https://opensource.ncsa.illinois.edu/jira/browse/CATS-884)
+- Added a new view to display Extractor Details (ExtractorInfo)
+  [CATS-892](https://opensource.ncsa.illinois.edu/jira/browse/CATS-892)
 - New API endpoints for proxying GET, POST, PUT, and DELETE requests through Clowder.
   [CATS-793](https://opensource.ncsa.illinois.edu/jira/browse/CATS-793)
   [CATS-889](https://opensource.ncsa.illinois.edu/jira/browse/CATS-889)
 
 ### Changed 
+- Displayed more ExtractorInfo in each Space's "Update Extractors" view.
+  [CATS-890](https://opensource.ncsa.illinois.edu/jira/browse/CATS-890)
 
 ### Fixed
 - In a private mode, a superadmin can now see datasets in a space that he/she is not part of.